+++ cli/deno.json
@@ -1,7 +1,8 @@
 {
   "tasks": {
     "run": "deno run --allow-read --allow-write --allow-env --allow-net --allow-run run.ts",
-    "install": "deno install --allow-read --allow-write --allow-env --allow-net --allow-run --config=deno.json --name ddbj-repository --force run.ts"
+    "install": "deno install --allow-read --allow-write --allow-env --allow-net --allow-run --config=deno.json --name ddbj-repository --force run.ts",
+    "test": "deno test --allow-read --allow-write --allow-env --allow-net --allow-run test.ts"
   },
   "imports": {
     "cliffy/": "https://deno.land/x/cliffy@v1.0.0-rc.3/",
